Lab for Week 1 Wednesday 7/8/2015
Income Tax Calculator which receives command line argument and, should
the command line argument not satisfy conditions, input from the user.
Method to retrieve seconds from time and use the input to display a
certain text depending on the current second retrieved.  (Proof of
Concept for switch integration?), Classic "FizzBuzz" exercise, method to
reverse all character in a string, method to reverse all strings in an
array and all the characters within that string (Proof of concept for
method usage)